requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
jinja2==3.1.6
python-dotenv==1.1.1
mysql-connector-python==9.3.0
langchain-core==0.3.66
langchain-community==0.3.26
langchain-openai==0.3.27
langchain==0.3.26
pydantic==2.11.7
starlette==0.46.2
